+++ packages/common/src/utils/arrays/index.ts
@@ -1,2 +1,3 @@
 export * from './utils';
 export * from './reducers';
+export * from './sorters';
