+++ packages/invoices/src/OutboundShipment/DetailView/SidePanel.tsx
@@ -15,7 +15,7 @@ import {
 } from '@openmsupply-client/common';
 import React, { FC } from 'react';
 import { Link } from 'react-router-dom';
-import { isInvoiceEditable } from '../utils';
+import { isInvoiceEditable } from '../../utils';
 import { OutboundShipment } from './types';
 
 interface SidePanelProps {
