+++ packages/invoices/src/OutboundShipment/DetailView/Toolbar.tsx
@@ -14,7 +14,7 @@ import {
 } from '@openmsupply-client/common';
 import { NameSearchInput } from '@openmsupply-client/system/src/Name';
 import { OutboundShipment, OutboundShipmentSummaryItem } from './types';
-import { isInvoiceEditable } from '../utils';
+import { isInvoiceEditable } from '../../utils';
 
 interface ToolbarProps {
   draft: OutboundShipment;
