+++ packages/invoices/src/OutboundShipment/DetailView/reducer.test.ts
@@ -1,5 +1,5 @@
 import { DocumentAction } from './../../../../common/src/hooks/useDocument/useDocument';
-import { flattenSummaryItems } from './../utils';
+import { flattenSummaryItems } from '../../utils';
 import { DocumentActionSet } from '@openmsupply-client/common';
 import { placeholderInvoice } from './index';
 import { reducer, OutboundShipmentStateShape, OutboundAction } from './reducer';
