+++ packages/invoices/src/utils.ts
@@ -4,7 +4,7 @@ import {
   OutboundShipment,
   OutboundShipmentRow,
   OutboundShipmentSummaryItem,
-} from './DetailView/types';
+} from './OutboundShipment/DetailView/types';
 
 export const outboundStatuses: OutboundShipmentStatus[] = [
   'DRAFT',
@@ -59,32 +59,6 @@ export const getStatusTranslation = (
 export const isInvoiceEditable = (outbound: OutboundShipment): boolean => {
   return outbound.status !== 'SHIPPED' && outbound.status !== 'DELIVERED';
 };
-
-const parseValue = (object: any, key: string) => {
-  const value = object[key];
-  if (typeof value === 'string') {
-    const valueAsNumber = Number.parseFloat(value);
-
-    if (!Number.isNaN(valueAsNumber)) return valueAsNumber;
-    return value.toUpperCase(); // ignore case
-  }
-  return value;
-};
-
-export const getDataSorter =
-  (sortKey: any, desc: boolean) => (a: any, b: any) => {
-    const valueA = parseValue(a, sortKey);
-    const valueB = parseValue(b, sortKey);
-
-    if (valueA < valueB) {
-      return desc ? 1 : -1;
-    }
-    if (valueA > valueB) {
-      return desc ? -1 : 1;
-    }
-
-    return 0;
-  };
 
 export const flattenSummaryItems = (
   summaryItems: OutboundShipmentSummaryItem[]
