+++ packages/mock-server/src/utils.ts
@@ -73,31 +73,6 @@ export const getFilter =
   (obj: T): boolean =>
     obj[key] === matchVal;
 
-const parseValue = <T>(object: T, key: keyof T) => {
-  const value = object[key];
-  if (typeof value === 'string') {
-    const valueAsNumber = Number.parseFloat(value);
-    if (!Number.isNaN(valueAsNumber)) return valueAsNumber;
-    return value.toUpperCase(); // ignore case
-  }
-  return value;
-};
-
-export const getDataSorter =
-  (sortKey: string, desc: boolean) => (a: any, b: any) => {
-    const valueA = parseValue(a, sortKey);
-    const valueB = parseValue(b, sortKey);
-
-    if (valueA < valueB) {
-      return desc ? 1 : -1;
-    }
-    if (valueA > valueB) {
-      return desc ? -1 : 1;
-    }
-
-    return 0;
-  };
-
 export const addRandomPercentageTo = ({
   value,
   min,
